Добавил класс applicationManager
Добавил класс applicationManager, своего рода усовершенствованный TestBase
