+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/tests/GroupCreationTest.java
@@ -1,17 +1,18 @@
 package ru.stqa.qajava.addressbook.tests;
 
-import org.testng.annotations.*;
+import org.testng.annotations.Test;
+import ru.stqa.qajava.addressbook.model.GroupData;
 
 
 public class GroupCreationTest extends TestBase{
 
   @Test
   public void testGroupCreation() throws Exception {
-    gotoGroupPage();
-    initGroupCreation();
-    fillGroupForm(new GroupData("te", "qerw", "er"));
-    submitGroupCreation();
-    reternToGroupPage();
+    app.gotoGroupPage();
+    app.initGroupCreation();
+    app.fillGroupForm(new GroupData("te", "qerw", "er"));
+    app.submitGroupCreation();
+    app.reternToGroupPage();
   }
 
 }
