+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/tests/TestBase.java
@@ -1,116 +1,21 @@
 package ru.stqa.qajava.addressbook.tests;
 
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
-
-import java.util.concurrent.TimeUnit;
-
-import static org.testng.Assert.fail;
+import ru.stqa.qajava.addressbook.appmanager.ApplicationManager;
 
 public class TestBase {
 
-  public WebDriver driver;
-  public String baseUrl;
-  public boolean acceptNextAlert = true;
-  public StringBuffer verificationErrors = new StringBuffer();
+  protected final ApplicationManager app = new ApplicationManager();// Делегирование
 
   @BeforeMethod
   public void setUp() throws Exception {
-    System.setProperty("webdriver.chrome.driver","/Users/alexandrkanishevskii/IdeaProjects/QAJava/addressbook-web-tests/src/test/resources/chromedriver");
-    driver = new ChromeDriver();
-    baseUrl = "https://www.katalon.com/";
-    driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-    driver.get("http://localhost/addressbook/group.php");
-    login("admin", "secret");
-  }
-
-  private void login(String username, String passord) {
-    driver.findElement(By.name("user")).click();
-    driver.findElement(By.name("user")).clear();
-    driver.findElement(By.name("user")).sendKeys(username);
-    driver.findElement(By.name("pass")).click();
-    driver.findElement(By.name("pass")).clear();
-    driver.findElement(By.name("pass")).sendKeys(passord);
-    driver.findElement(By.xpath("(.//*[normalize-space(text()) and normalize-space(.)='Password:'])[1]/following::input[2]")).click();
-  }
-
-  protected void reternToGroupPage() {
-    driver.findElement(By.linkText("group page")).click();
-  }
-
-  protected void submitGroupCreation() {
-    driver.findElement(By.name("submit")).click();
-  }
-
-  protected void fillGroupForm(GroupData groupData) {
-    driver.findElement(By.name("group_name")).click();
-    driver.findElement(By.name("group_name")).clear();
-    driver.findElement(By.name("group_name")).sendKeys(groupData.getGroupName());
-    driver.findElement(By.name("group_header")).click();
-    driver.findElement(By.name("group_header")).clear();
-    driver.findElement(By.name("group_header")).sendKeys(groupData.getGropupHeader());
-    driver.findElement(By.name("group_footer")).click();
-    driver.findElement(By.name("group_footer")).clear();
-    driver.findElement(By.name("group_footer")).sendKeys(groupData.getGroupFooter());
-  }
-
-  protected void initGroupCreation() {
-    driver.findElement(By.name("new")).click();
-  }
-
-  protected void gotoGroupPage() {
-    driver.findElement(By.linkText("groups")).click();
+    app.init();
   }
 
   @AfterMethod
   public void tearDown() throws Exception {
-    driver.quit();
-    String verificationErrorString = verificationErrors.toString();
-    if (!"".equals(verificationErrorString)) {
-      fail(verificationErrorString);
-    }
+    app.stop();
   }
 
-  private boolean isElementPresent(By by) {
-    try {
-      driver.findElement(by);
-      return true;
-    } catch (NoSuchElementException e) {
-      return false;
-    }
-  }
-
-  private boolean isAlertPresent() {
-    try {
-      driver.switchTo().alert();
-      return true;
-    } catch (NoAlertPresentException e) {
-      return false;
-    }
-  }
-
-  private String closeAlertAndGetItsText() {
-    try {
-      Alert alert = driver.switchTo().alert();
-      String alertText = alert.getText();
-      if (acceptNextAlert) {
-        alert.accept();
-      } else {
-        alert.dismiss();
-      }
-      return alertText;
-    } finally {
-      acceptNextAlert = true;
-    }
-  }
-
-  protected void deleteSelectedGroups() {
-    driver.findElement(By.name("delete")).click();
-  }
-
-  protected void selectGroup() {
-    driver.findElement(By.xpath("(.//*[normalize-space(text()) and normalize-space(.)='test1'])[1]/input[1]")).click();
-  }
 }
